added watchdog service
* added CheckRestartSerivice that checks the Bizplay server for the need
to reboot

* since a reboot cannot be performed by 'normal' apps on a non rooted
device the app itself is restarted

* added colaboration between the MainActivity and CheckRestartService on
the base of a binding of the new IserviceCallBacks Interface that is
implemented by the MainActivity. The MainActivity starts the service so
it runs in the same thread and the inter thread communicaton that is
adviced for loosly coupled services is not needed.

* the new service is added to AndroidManifest, an unused service has
been removed from it.